pyproject.toml
[project]
name = "rag-evaluation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "faiss-cpu>=1.9.0",
    "httpx>=0.28.0",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-milvus>=0.2.0",
    "langchain-openai>=1.1.12",
    "langchain-openrouter>=0.2.1",
    "langchain-postgres>=0.0.13",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=1.1.1",
    "langchain-weaviate>=0.0.2",
    "langchainhub>=0.1.21",
    "lancedb>=0.20.0",
    "numpy>=1.26.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.0",
    "pycryptodome>=3.23.0",
    "pymupdf>=1.25.0",
    "pymupdf4llm>=0.0.17",
    "pypdf>=6.10.0",
    "tqdm>=4.67.3",
    "weaviate-client>=4.10.0",
]
